@@ -1,52 +1,52 @@
-from setuptools import setup, find_packages
+from setuptools import setup, find_namespace_packages
 from os import path
 
 requires = [
-    "asyncpg",
-    "click-plugins",
+    "asyncpg~=0.25",
+    "click-plugins~=1.0",
     # click has been known to publish non-backwards compatible minors in the past (removed deprecated code in 8.1.0)
     "click>=8.0,<8.2",
-    "colorlog",
-    "cookiecutter",
-    "cryptography",
+    "colorlog~=6.0",
+    "cookiecutter>=1,<3",
+    "crontab~=0.23",
+    "cryptography>=36,<38",
     # docstring-parser has been known to publish non-backwards compatible minors in the past
-    "docstring-parser>=0.10,<0.14",
-    "email-validator",
-    "execnet",
-    "importlib_metadata",
-    "jinja2",
-    "more-itertools",
-    "netifaces",
-    "packaging",
+    "docstring-parser>=0.10,<0.15",
+    "email-validator~=1.0",
+    "execnet~=1.0",
+    "importlib_metadata~=4.0",
+    "jinja2~=3.0",
+    "more-itertools~=8.0",
+    "netifaces~=0.11",
+    "packaging~=21.3",
     # pip>=21.3 required for editable pyproject.toml + setup.cfg based install support
     "pip>=21.3",
-    "ply",
+    "ply~=3.0",
     # Exclude pre-release due to https://github.com/samuelcolvin/pydantic/issues/3546
-    "pydantic!=1.9.0a1",
-    "pyformance",
-    "PyJWT",
-    "python-dateutil",
-    "pyyaml",
-    "texttable",
-    "tornado>=6.0",
-    "typing_inspect",
-    "build",
-    "ruamel.yaml",
+    "pydantic~=1.0,!=1.9.0a1",
+    "pyformance~=0.4",
+    "PyJWT~=2.0",
+    "python-dateutil~=2.0",
+    "pyyaml~=6.0",
+    "texttable~=1.0",
+    "tornado~=6.0",
+    "typing_inspect~=0.7",
+    "build~=0.7",
+    "ruamel.yaml~=0.17",
+    "toml~=0.10 ",
 ]
 
-# Package a dummy extensions so that the namespace package for extensions is not empty
-namespace_packages = ["inmanta_ext.core", "inmanta_plugins.1"]
 
 # read the contents of your README file
 this_directory = path.abspath(path.dirname(__file__))
 with open(path.join(this_directory, "README.md"), encoding="utf-8") as f:
     long_description = f.read()
 
-version = "6.1.0"
+version = "6.2.0"
 
 setup(
     version=version,
-    python_requires=">=3.6",  # also update classifiers
+    python_requires=">=3.9",  # also update classifiers
     # Meta data
     name="inmanta-core",
     description="Inmanta deployment tool",
@@ -63,9 +63,6 @@ setup(
         "Operating System :: POSIX :: Linux",
         "Topic :: System :: Systems Administration",
         "Topic :: Utilities",
-        "Programming Language :: Python :: 3.6",
-        "Programming Language :: Python :: 3.7",
-        "Programming Language :: Python :: 3.8",
         "Programming Language :: Python :: 3.9",
     ],
     keywords="orchestrator orchestration configurationmanagement",
@@ -75,9 +72,10 @@ setup(
     },
     # Packaging
     package_dir={"": "src"},
-    packages=find_packages("src") + namespace_packages,
+    # All data files should be treated as namespace package according to
+    # https://setuptools.pypa.io/en/latest/userguide/datafiles.html#subdirectory-for-data-files
+    packages=find_namespace_packages(where="src"),
     # https://www.python.org/dev/peps/pep-0561/#packaging-type-information
-    package_data={"": ["misc/*", "docs/*"], "inmanta": ["py.typed"]},
     zip_safe=False,
     include_package_data=True,
     install_requires=requires,
