@@ -16,4 +16,14 @@
     Contact: code@inmanta.com
 """
 
-COMPILER_VERSION = "2022.2"
+COMPILER_VERSION = "2022.3"
+RUNNING_TESTS = False
+"""
+    This is enabled/disabled by the test suite when tests are run.
+    This variable is used to disable certain features that shouldn't run during tests.
+"""
+
+if __name__ == "__main__":
+    import inmanta.app
+
+    inmanta.app.app()
