@@ -19,7 +19,7 @@
 import traceback
 from abc import abstractmethod
 from functools import lru_cache
-from typing import Dict, FrozenSet, Iterator, List, Optional, Sequence, Tuple, Union  # noqa: F401
+from typing import Dict, List, Optional, Union
 
 from inmanta.ast import export
 from inmanta.stable_api import stable_api
@@ -32,11 +32,8 @@ except ImportError:
 
 
 if TYPE_CHECKING:
-    import inmanta.ast.statements  # noqa: F401
     from inmanta.ast.attribute import Attribute  # noqa: F401
-    from inmanta.ast.blocks import BasicBlock  # noqa: F401
-    from inmanta.ast.entity import Entity  # noqa: F401
-    from inmanta.ast.statements import AssignStatement, Statement  # noqa: F401
+    from inmanta.ast.statements import Statement  # noqa: F401
     from inmanta.ast.statements.define import DefineEntity, DefineImport  # noqa: F401
     from inmanta.ast.type import NamedType, Type  # noqa: F401
     from inmanta.compiler import Compiler
@@ -149,11 +146,10 @@ class Range(Location):
 
 
 class Locatable(object):
-
-    _location: Location
+    __slots__ = ("_location",)
 
     def __init__(self) -> None:
-        self._location = None
+        self._location: Optional[Location] = None
 
     def set_location(self, location: Location) -> None:
         assert location is not None and location.lnr > 0
@@ -162,6 +158,12 @@ class Locatable(object):
     def get_location(self) -> Location:
         assert self._location is not None
         return self._location
+
+    def copy_location(self, other: "Locatable") -> None:
+        """
+        Copy the location of this locatable to the given locatable
+        """
+        other.set_location(self.location)
 
     location = property(get_location, set_location)
 
@@ -243,6 +245,8 @@ class AttributeReferenceAnchor(Anchor):
 
 
 class Namespaced(Locatable):
+    __slots__ = ()
+
     @abstractmethod
     def get_namespace(self) -> "Namespace":
         raise NotImplementedError()
