@@ -69,6 +69,9 @@ class InmantaBootloader(object):
     - starting the server and its slices in the correct order
     """
 
+    # Cache field for available extensions
+    AVAILABLE_EXTENSIONS: Optional[Dict[str, str]] = None
+
     def __init__(self) -> None:
         self.restserver = Server()
         self.started = False
@@ -99,6 +102,24 @@ class InmantaBootloader(object):
         if self.feature_manager is not None:
             self.feature_manager.stop()
 
+    @classmethod
+    def get_available_extensions(cls) -> Dict[str, str]:
+        """
+        Returns a dictionary of with all available inmanta extensions.
+        The key contains the name of the extension and the value the fully qualified path to the python package.
+        """
+        if cls.AVAILABLE_EXTENSIONS is None:
+            try:
+                inmanta_ext = importlib.import_module(EXTENSION_NAMESPACE)
+            except ModuleNotFoundError:
+                # This only happens when a test case creates and activates a new venv
+                return {}
+            else:
+                cls.AVAILABLE_EXTENSIONS = {
+                    name[len(EXTENSION_NAMESPACE) + 1 :]: name for finder, name, ispkg in iter_namespace(inmanta_ext)
+                }
+        return dict(cls.AVAILABLE_EXTENSIONS)
+
     # Extension loading Phase I: from start to setup functions collected
     def _discover_plugin_packages(self) -> List[str]:
         """Discover all packages that are defined in the inmanta_ext namespace package. Filter available extensions based on
@@ -106,9 +127,7 @@ class InmantaBootloader(object):
 
         :return: A list of all subpackages defined in inmanta_ext
         """
-        inmanta_ext = importlib.import_module(EXTENSION_NAMESPACE)
-        available = {name[len(EXTENSION_NAMESPACE) + 1 :]: name for finder, name, ispkg in iter_namespace(inmanta_ext)}
-
+        available = self.get_available_extensions()
         LOGGER.info("Discovered extensions: %s", ", ".join(available.keys()))
 
         extensions = []
