@@ -197,8 +197,9 @@ server_version_to_keep = Option(
     "server",
     "available-versions-to-keep",
     10,
-    """On boot and at regular intervals the server will purge older versions.
-                                   This is the number of most recent versions to keep available.""",
+    """[DEPRECATED: use AVAILABLE_VERSIONS_TO_KEEP environment setting]
+                                On boot and at regular intervals the server will purge older versions.
+                                This is the number of most recent versions to keep available.""",
     is_int,
 )
 
