@@ -5,21 +5,20 @@ from asyncpg.cursor import CursorFactory
 from asyncpg.prepared_stmt import PreparedStatement
 from asyncpg.protocol import Record
 
-from . import connect_utils, types
+from . import connect_utils
 from asyncpg.transaction import Transaction
 from collections import namedtuple
 from typing import Any, Optional, List, AsyncIterator
 
+
 class Connection:
-    def __init__(self, protocol: Any, transport: Any, loop: Any, addr: None, config: connect_utils._ClientConfiguration, params: connect_utils._ConnectionParameters) -> None: ...
+    def __init__(self, protocol: Any, transport: Any, loop: Any, addr: None, config: connect_utils._ClientConfiguration,
+                 params: connect_utils._ConnectionParameters) -> None: ...
     def __del__(self) -> None: ...
     async def add_listener(self, channel: Any, callback: Any) -> None: ...
     async def remove_listener(self, channel: Any, callback: Any) -> None: ...
     def add_log_listener(self, callback: Any) -> None: ...
     def remove_log_listener(self, callback: Any) -> None: ...
-    def get_server_pid(self) -> int: ...
-    def get_server_version(self) -> types.ServerVersion: ...
-    def get_settings(self) -> object: ... # The real type seems to be generated and does not exist in the python codebase
     def transaction(self, *, isolation: str = ..., readonly: bool = ..., deferrable: bool = ...) -> Transaction: ...
     def is_in_transaction(self) -> bool: ...
     async def execute(self, query: str, *args: Any, timeout: float=...) -> str: ...
@@ -43,5 +42,22 @@ class Connection:
     async def reload_schema_state(self) -> None: ...
 
 async def connect(dsn: Optional[Any] = ..., *, host: Optional[Any] = ..., port: Optional[Any] = ..., user: Optional[Any] = ..., password: Optional[Any] = ..., passfile: Optional[Any] = ..., database: Optional[Any] = ..., loop: Optional[Any] = ..., timeout: int = ..., statement_cache_size: int = ..., max_cached_statement_lifetime: int = ..., max_cacheable_statement_size: Any = ..., command_timeout: Optional[Any] = ..., ssl: Optional[Any] = ..., connection_class: Any = ..., server_settings: Optional[Any] = ...) -> Connection: ...
+
+class _StatementCacheEntry:
+    def __init__(self, cache: Any, query: Any, statement: Any) -> None: ...
+
+class _StatementCache:
+    def __init__(self, loop: Any, max_size: Any, on_remove: Any, max_lifetime: Any) -> None: ...
+    def set_max_size(self, new_size: Any) -> None: ...
+    def set_max_lifetime(self, new_lifetime: Any) -> None: ...
+    def put(self, query: Any, statement: Any) -> None: ...
+    def clear(self) -> None: ...
+
+class _Atomic:
+    def __init__(self) -> None: ...
+    def __enter__(self) -> None: ...
+    def __exit__(self, t: Any, e: Any, tb: Any) -> None: ...
+
+class _ConnectionProxy: ...
 
 ServerCapabilities = namedtuple('ServerCapabilities', ['advisory_locks', 'notifications', 'plpgsql', 'sql_reset', 'sql_close_all'])
