@@ -2,14 +2,11 @@
 #
 # NOTE: This dynamically typed stub was automatically generated by stubgen.
 
-from asyncpg.protocol import Record
-from . import connresource
-from typing import Any, List, Optional
+from . import connresource, Record
+from typing import Any, Optional, List
 
 class CursorFactory(connresource.ConnectionResource):
     def __init__(self, connection: Any, query: Any, state: Any, args: Any, prefetch: Any, timeout: Any) -> None: ...
-    def __aiter__(self): ...
-    def __await__(self): ...
     def __del__(self) -> None: ...
 
 class BaseCursor(connresource.ConnectionResource):
@@ -18,8 +15,6 @@ class BaseCursor(connresource.ConnectionResource):
 
 class CursorIterator(BaseCursor):
     def __init__(self, connection: Any, query: Any, state: Any, args: Any, prefetch: Any, timeout: Any) -> None: ...
-    def __aiter__(self): ...
-    async def __anext__(self): ...
 
 class Cursor(BaseCursor):
     async def fetch(self, n: Any, *, timeout: Optional[Any] = ...) -> List[Record]: ...
