@@ -2,19 +2,14 @@
 #
 # NOTE: This dynamically typed stub was automatically generated by stubgen.
 
-from typing import Any, Optional
+from typing import Any, Optional, Type
 
-class PostgresMessageMeta(type):
-    def __new__(mcls: Any, name: Any, bases: Any, dct: Any): ...
-    @classmethod
-    def get_message_class_for_sqlstate(mcls: Any, code: Any): ...
 
-class PostgresMessage(metaclass=PostgresMessageMeta):
-    def as_dict(self): ...
+class PostgresMessageMeta(type): ...
 
-class PostgresError(PostgresMessage, Exception):
-    @classmethod
-    def new(cls, fields: Any, query: Optional[Any] = ...): ...
+class PostgresMessage(metaclass=PostgresMessageMeta): ...
+
+class PostgresError(PostgresMessage, Exception): ...
 
 class FatalPostgresError(PostgresError): ...
 class UnknownPostgresError(FatalPostgresError): ...
@@ -43,5 +38,3 @@ class OutdatedSchemaCacheError(InternalClientError):
 
 class PostgresLogMessage(PostgresMessage):
     def __setattr__(self, name: Any, val: Any) -> None: ...
-    @classmethod
-    def new(cls, fields: Any, query: Optional[Any] = ...): ...
