@@ -1,42 +1,40 @@
 # Stubs for asyncpg.pool (Python 3)
 #
 # NOTE: This dynamically typed stub was automatically generated by stubgen.
-from . import connection
+from asyncpg import Connection
+
+from . import connection, Record
 from typing import Any, Optional, AsyncContextManager, List
 
-
-class PoolConnectionHolder:
-    _con: connection.Connection
-
-
 class Pool:
+    _closing: bool
+    _closed: bool
     _maxsize: int
-    _holders: List[PoolConnectionHolder]
-
     def __init__(self, *connect_args: Any, min_size: Any, max_size: Any, max_queries: Any, max_inactive_connection_lifetime: Any, setup: Any, init: Any, loop: Any, connection_class: Any, **connect_kwargs: Any) -> None: ...
-    async def _async__init__(self): ...
     def set_connect_args(self, dsn: Optional[Any] = ..., **connect_kwargs: Any) -> None: ...
     async def execute(self, query: str, *args: Any, timeout: float=...) -> str: ...
     async def executemany(self, command: str, args: Any, *, timeout: float=...) -> Any: ...
-    async def fetch(self, query: Any, *args: Any, timeout: Any=...) -> list: ...
-    async def fetchval(self, query: Any, *args: Any, column: int = ..., timeout: Optional[Any] = ...): ...
-    async def fetchrow(self, query: Any, *args: Any, timeout: Optional[Any] = ...): ...
-    def acquire(self, *, timeout: float = ...) -> AsyncContextManager[connection.Connection]: ...
-    async def release(self, connection: Any, *, timeout: Optional[Any] = ...): ...
+    async def fetch(self, query: Any, *args: Any, timeout: Any=...) -> List[Record]: ...
+    async def fetchval(self, query: Any, *args: Any, column: int = ..., timeout: Optional[Any] = ...) -> object: ...
+    async def fetchrow(self, query: Any, *args: Any, timeout: Optional[Any] = ...) -> Record: ...
+    def acquire(self, *, timeout: float = ...) -> AsyncContextManager[Connection]: ...
+    async def release(self, connection: Any, *, timeout: Optional[Any] = ...) -> None: ...
     async def close(self) -> None: ...
+    def get_max_size(self) -> int: ...
+    def get_size(self) -> int: ...
+    def get_idle_size(self) -> int: ...
     def terminate(self) -> None: ...
     async def expire_connections(self) -> None: ...
-    def __await__(self): ...
-    async def __aenter__(self): ...
     async def __aexit__(self, *exc: Any) -> None: ...
-
 
 class PoolAcquireContext:
-    pool: Pool
+    pool: Any
+    timeout: Any
+    connection: Any
+    done: bool
     def __init__(self, pool: Pool, timeout: int) -> None: ...
-    async def __aenter__(self) -> connection.Connection: ...
-    async def __aexit__(self, *exc: Any) -> None: ...
-    def __await__(self) -> connection.Connection: ...
+    async def __aenter__(self) -> Connection: ...
+    async def __aexit__(self, *exc: object) -> None: ...
+    def __await__(self) -> AsyncContextManager[Connection]: ...
 
-
-def create_pool(dsn: Optional[Any] = ..., *, min_size: int = ..., max_size: int = ..., max_queries: int = ..., max_inactive_connection_lifetime: float = ..., setup: Optional[Any] = ..., init: Optional[Any] = ..., loop: Optional[Any] = ..., connection_class: Any = ..., **connect_kwargs: Any): ...
+def create_pool(dsn: Optional[Any] = ..., *, min_size: int = ..., max_size: int = ..., max_queries: int = ..., max_inactive_connection_lifetime: float = ..., setup: Optional[Any] = ..., init: Optional[Any] = ..., loop: Optional[Any] = ..., connection_class: Any = ..., **connect_kwargs: Any) -> Pool: ...
