@@ -2,18 +2,16 @@
 #
 # NOTE: This dynamically typed stub was automatically generated by stubgen.
 
-from . import connresource, cursor
-from typing import Any, Optional
+from . import connresource, cursor, Record
+from typing import Any, Optional, List
+
 
 class PreparedStatement(connresource.ConnectionResource):
     def __init__(self, connection: Any, query: Any, state: Any) -> None: ...
     def get_query(self) -> str: ...
     def get_statusmsg(self) -> str: ...
-    def get_parameters(self): ...
-    def get_attributes(self): ...
     def cursor(self, *args: Any, prefetch: Any=..., timeout: Any=...) -> cursor.CursorFactory: ...
-    async def explain(self, *args: Any, analyze: bool = ...): ...
-    async def fetch(self, *args: Any, timeout: Optional[Any] = ...): ...
-    async def fetchval(self, *args: Any, column: int = ..., timeout: Optional[Any] = ...): ...
-    async def fetchrow(self, *args: Any, timeout: Optional[Any] = ...): ...
+    async def fetch(self, *args: Any, timeout: Optional[Any] = ...) -> List[Record]: ...
+    async def fetchval(self, *args: Any, column: int = ..., timeout: Optional[Any] = ...) -> object: ...
+    async def fetchrow(self, *args: Any, timeout: Optional[Any] = ...) -> Record: ...
     def __del__(self) -> None: ...
