@@ -1,17 +1,11 @@
 # Stubs for asyncpg.protocol (Python 3)
 #
 # NOTE: This dynamically typed stub was automatically generated by stubgen.
-from typing import Union, Any, Iterator, Tuple
+from typing import Union, Any, Iterator, Tuple, Mapping, Optional, ValuesView, AbstractSet
 
-
-class Record(object):
-
+class Record(Mapping[str, object]):
     def __len__(self) -> int: ...
-    def __getitem__(self, item: Union[str, int]) -> Any: ...
-    def __contains__(self, item: str) -> bool: ...
-    def __iter__(self) -> Iterator[Any]: ...
-    def get(self, name:str, default: Any) -> Any: ...
-    def values(self) -> Iterator[Any]: ...
-    def keys(self) -> Iterator[str]:...
-    def items(self) -> Iterator[Tuple[str, Any]]:...
+    def __getitem__(self, item: Union[str, int]) -> object: ...
+    def __contains__(self, item: object) -> bool: ...
+    def __iter__(self) -> Iterator[str]: ...
 
