@@ -12,7 +12,25 @@ Attribute = namedtuple('Attribute', ['name', 'type'])
 
 ServerVersion = namedtuple('ServerVersion', ['major', 'minor', 'micro', 'releaselevel', 'serial'])
 
-class Range: ...
+# class Range:
+#     def __init__(self, lower: Optional[Any] = ..., upper: Optional[Any] = ..., *, lower_inc: bool = ..., upper_inc: bool = ..., empty: bool = ...) -> None: ...
+#     @property
+#     def lower(self): ...
+#     @property
+#     def lower_inc(self): ...
+#     @property
+#     def lower_inf(self): ...
+#     @property
+#     def upper(self): ...
+#     @property
+#     def upper_inc(self): ...
+#     @property
+#     def upper_inf(self): ...
+#     @property
+#     def isempty(self): ...
+#     def __bool__(self): ...
+#     def __eq__(self, other: Any): ...
+#     def __hash__(self): ...
 
 # Names in __all__ with no definition:
 #   Attribute
