@@ -236,7 +236,8 @@ def test_bad_deref(snippetcompiler):
 h = std::Host(name="test", os=std::linux)
 std::print(h.name.test)
 """,
-        "can not get a attribute test, test not an entity (reported in h.name.test ({dir}/main.cf:3))",
+        "can not get attribute test, h.name is not an entity but a str with value 'test'"
+        " (reported in h.name.test ({dir}/main.cf:3))",
     )
 
 
