@@ -39,7 +39,8 @@ def check_result(result):
 async def test_dump_db(server, client, postgres_db, database_name):
     """
     Note: remove following line from the dump: SELECT pg_catalog.set_config('search_path', '', false);
-
+    The line 'SET default_table_access_method = heap;' should also be removed.
+    Also be careful that the IDE doesn't remove trailing whitespace from the dump file.
     """
 
     if False:
