@@ -26,11 +26,14 @@ import pytest
 from pkg_resources import Requirement
 
 from inmanta import plugins
+from inmanta.ast import CompilerException
 from inmanta.compiler.config import feature_compiler_cache
-from inmanta.env import LocalPackagePath
+from inmanta.const import CF_CACHE_DIR
+from inmanta.env import ConflictingRequirements, LocalPackagePath, process_env
 from inmanta.module import (
     DummyProject,
     InmantaModuleRequirement,
+    InvalidModuleException,
     ModuleLoadingException,
     ModuleNotFoundException,
     ModuleV1,
@@ -38,14 +41,15 @@ from inmanta.module import (
     ModuleV2Source,
     Project,
 )
-from inmanta.moduletool import ModuleConverter
-from utils import PipIndex, module_from_template, v1_module_from_template
+from inmanta.moduletool import ModuleConverter, ModuleTool
+from packaging.version import Version
+from utils import PipIndex, create_python_package, module_from_template, v1_module_from_template
 
 
 @pytest.mark.parametrize_any("editable_install", [True, False])
 def test_v2_module_loading(editable_install: bool, tmpdir: py.path.local, snippetcompiler, capsys, modules_v2_dir: str) -> None:
     # Work around caching problem in venv
-    feature_compiler_cache.set("False")
+    feature_compiler_cache.set("True")
     # Disable modules_dir
     snippetcompiler.modules_dir = None
 
@@ -67,6 +71,10 @@ def test_v2_module_loading(editable_install: bool, tmpdir: py.path.local, snippe
 
     snippetcompiler.do_export()
     assert "Hello world" in capsys.readouterr().out
+
+    # Make sure the cache files are created
+    cache_folder = os.path.join(snippetcompiler.project_dir, CF_CACHE_DIR)
+    assert len(os.listdir(cache_folder)) > 0
 
 
 def test_v1_and_v2_module_installed_simultaneously(
@@ -526,3 +534,258 @@ def test_module_has_v2_requirements_on_non_imported_module(snippetcompiler, loca
     )
     project.load_module_recursive()
     assert "minimalv2module" not in project.modules
+
+
+@pytest.mark.slowtest
+def test_module_v2_load_installed_without_required(snippetcompiler_clean, local_module_package_index: str) -> None:
+    """
+    Verify that module loading works as expected for a v2 module installed in the environment while it is not in any of the
+    requirements files. This is relevant for correct working of pytest-inmanta's auto-generated projects.
+    """
+    # set up venv
+    snippetcompiler_clean.setup_for_snippet("", autostd=False)
+    process_env.install_from_index(
+        [ModuleV2Source.get_python_package_requirement(InmantaModuleRequirement.parse("elaboratev2module"))],
+        index_urls=[local_module_package_index],
+    )
+
+    # set up project: import submodule only, don't add the module to requirements
+    project: Project = snippetcompiler_clean.setup_for_snippet("import elaboratev2module::nested::sub", autostd=False)
+    assert {tup[0] for tup in project.load_module_recursive()} == {
+        "elaboratev2module",
+        "elaboratev2module::nested",
+        "elaboratev2module::nested::sub",
+    }
+
+
+@pytest.mark.slowtest
+def test_project_requirements_dont_overwrite_core_requirements_source(
+    snippetcompiler_clean,
+    local_module_package_index: str,
+    modules_v2_dir: str,
+    tmpdir: py.path.local,
+) -> None:
+    """
+    A project has a requirement that is also a requirement of core
+    but with another version. The requirements of core should not be
+    overwritten. The module gets installed from source
+    """
+    if "inmanta-core" in process_env.get_installed_packages(only_editable=True):
+        pytest.skip(
+            "This test would fail if it runs against an inmanta-core installed in editable mode, because the build tag "
+            "on the development branch is set to .dev0. The inmanta package protection feature would make pip "
+            "install a non-editable version of the same package. But no version with build tag .dev0 exists on the python "
+            "package repository."
+        )
+
+    # Create the module
+    module_name: str = "minimalv2module"
+    module_path: str = str(tmpdir.join(module_name))
+    module_from_template(
+        os.path.join(modules_v2_dir, module_name), module_path, new_requirements=[Requirement.parse("Jinja2==2.11.3")]
+    )
+
+    # Activate the snippetcompiler venv
+    project: Project = snippetcompiler_clean.setup_for_snippet("")
+    active_env = project.virtualenv
+    jinja2_version_before = active_env.get_installed_packages()["Jinja2"].base_version
+
+    # Install the module
+    with pytest.raises(InvalidModuleException) as e:
+        ModuleTool().install(editable=False, path=module_path)
+
+    assert ("Module installation failed due to conflicting dependencies") in str(e.value.msg)
+
+    jinja2_version_after = active_env.get_installed_packages()["Jinja2"].base_version
+    assert jinja2_version_before == jinja2_version_after
+
+
+@pytest.mark.slowtest
+def test_project_requirements_dont_overwrite_core_requirements_index(
+    snippetcompiler_clean,
+    modules_v2_dir: str,
+    tmpdir: py.path.local,
+) -> None:
+    """
+    A module from index has a requirement that is also a requirement of core
+    but with another version. The requirements of core should not be
+    overwritten. The module gets installed from index.
+    """
+    if "inmanta-core" in process_env.get_installed_packages(only_editable=True):
+        pytest.skip(
+            "This test would fail if it runs against an inmanta-core installed in editable mode, because the build tag "
+            "on the development branch is set to .dev0. The inmanta package protection feature would make pip "
+            "install a non-editable version of the same package. But no version with build tag .dev0 exists on the python "
+            "package repository."
+        )
+    # Create the module
+    module_name: str = "minimalv2module"
+    module_path: str = str(tmpdir.join(module_name))
+    index: PipIndex = PipIndex(artifact_dir=os.path.join(str(tmpdir), ".custom-index"))
+    module_from_template(
+        os.path.join(modules_v2_dir, module_name),
+        module_path,
+        new_requirements=[Requirement.parse("Jinja2==2.11.3")],
+        publish_index=index,
+    )
+
+    # Setup project
+    project: Project = snippetcompiler_clean.setup_for_snippet(
+        "",
+        install_project=False,
+        python_package_sources=[index.url, "https://pypi.org/simple"],
+        python_requires=[ModuleV2Source.get_python_package_requirement(InmantaModuleRequirement.parse(module_name))],
+        autostd=False,
+    )
+
+    active_env = project.virtualenv
+    jinja2_version_before = active_env.get_installed_packages()["Jinja2"].base_version
+
+    # Install project
+    with pytest.raises(ConflictingRequirements):
+        project.install_modules()
+
+    jinja2_version_after = active_env.get_installed_packages()["Jinja2"].base_version
+    assert jinja2_version_before == jinja2_version_after
+
+
+@pytest.mark.slowtest
+@pytest.mark.parametrize("strict_deps_check", [True, False, None])
+def test_module_conflicting_dependencies_with_v2_modules(
+    snippetcompiler_clean,
+    modules_v2_dir: str,
+    tmpdir: py.path.local,
+    caplog,
+    strict_deps_check: Optional[bool],
+) -> None:
+    """
+    Show an error message when installing a module that breaks the dependencies
+    of another one. minimalv2module depends on y~=1.0.0 which requires x~=1.0.0.
+    after the install of minimalv2module we try to install minimalv2module2 which
+    requires x~=2.0.0. The y~=1.0.0 requirement is now broken as python package x
+    has now version 2.0.0 and y needs 1.0.0.
+    """
+    index: PipIndex = PipIndex(artifact_dir=os.path.join(str(tmpdir), ".custom-index"))
+
+    # Create a python package x with version 1.0.0
+    create_python_package("x", Version("1.0.0"), str(tmpdir.join("x-1.0.0")), publish_index=index)
+
+    # Create a python package x with version 2.0.0
+    create_python_package("x", Version("2.0.0"), str(tmpdir.join("x-2.0.0")), publish_index=index)
+
+    # Create a python package y with version 1.0.0 that depends on x~=1.0.0
+    create_python_package(
+        "y", Version("1.0.0"), str(tmpdir.join("y-1.0.0")), requirements=[Requirement.parse("x~=1.0.0")], publish_index=index
+    )
+
+    # Create the first module
+    module_name1: str = "minimalv2module"
+    module_path1: str = str(tmpdir.join(module_name1))
+    module_from_template(
+        os.path.join(modules_v2_dir, module_name1),
+        module_path1,
+        new_requirements=[Requirement.parse("y~=1.0.0")],
+        publish_index=index,
+    )
+
+    # Create the second module
+    module_name2: str = "minimalv2module2"
+    module_path2: str = str(tmpdir.join(module_name2))
+    module_from_template(
+        os.path.join(modules_v2_dir, "minimalv2module"),
+        module_path2,
+        new_name="minimalv2module2",
+        new_requirements=[Requirement.parse("x~=2.0.0")],
+        publish_index=index,
+    )
+
+    req1 = ModuleV2Source.get_python_package_requirement(InmantaModuleRequirement.parse(module_name1))
+    req2 = ModuleV2Source.get_python_package_requirement(InmantaModuleRequirement.parse(module_name2))
+    # Setup project
+    project: Project = snippetcompiler_clean.setup_for_snippet(
+        "",
+        install_project=False,
+        python_package_sources=[index.url],
+        python_requires=[req1, req2],
+        autostd=False,
+        strict_deps_check=strict_deps_check,
+    )
+
+    # Install project
+    msg: str = "Incompatibility between constraint x"
+    if strict_deps_check:
+        with pytest.raises(ConflictingRequirements) as e:
+            project.install_modules()
+        assert msg in e.value.get_message()
+    else:
+        # The version conflict is present in a transitive dependency, so without strict_deps_check enabled,
+        # only a warning message will be logged.
+        with caplog.at_level(logging.WARNING):
+            caplog.clear()
+            project.install_modules()
+            assert msg in caplog.text
+
+
+@pytest.mark.slowtest
+@pytest.mark.parametrize("strict_deps_check", [True, False, None])
+def test_module_conflicting_dependencies_with_v1_module(
+    snippetcompiler_clean,
+    modules_dir: str,
+    modules_v2_dir: str,
+    tmpdir: py.path.local,
+    strict_deps_check: Optional[bool],
+) -> None:
+    """
+    Show an error message when installing a module that breaks the dependencies
+    of another one. modulev1 depends on y~=1.0.0.
+    after the install of modulev1 we try to install minimalv2module2 which
+    requires y~=2.0.0. those 2 requirements conflict which each other.
+    """
+    index: PipIndex = PipIndex(artifact_dir=os.path.join(str(tmpdir), ".custom-index"))
+    # Create a python package x with version 1.0.0
+    create_python_package("y", Version("1.0.0"), str(tmpdir.join("y-1.0.0")), publish_index=index)
+
+    # Create a python package x with version 2.0.0
+    create_python_package("y", Version("2.0.0"), str(tmpdir.join("y-2.0.0")), publish_index=index)
+
+    # Create the first module
+    module_name1: str = "minimalv1module"
+    module_path1: str = str(tmpdir.join("modulev1"))
+    v1_module_from_template(
+        os.path.join(modules_dir, module_name1),
+        module_path1,
+        new_name="modulev1",
+        new_requirements=[Requirement.parse("y~=1.0.0")],
+    )
+
+    # Create the second module
+    module_name2: str = "minimalv2module"
+    module_path2: str = str(tmpdir.join(module_name2))
+    module_from_template(
+        os.path.join(modules_v2_dir, module_name2),
+        module_path2,
+        new_requirements=[Requirement.parse("y~=2.0.0")],
+        publish_index=index,
+    )
+
+    req = ModuleV2Source.get_python_package_requirement(InmantaModuleRequirement.parse(module_name2))
+
+    # Setup project
+    project: Project = snippetcompiler_clean.setup_for_snippet(
+        "import modulev1",
+        install_project=False,
+        python_package_sources=[index.url],
+        python_requires=[req],
+        autostd=False,
+        add_to_module_path=[str(tmpdir)],
+        strict_deps_check=strict_deps_check,
+    )
+
+    # Install project
+    with pytest.raises(CompilerException) as e:
+        # The version conflict is present in a direct dependency, so this always results in an error.
+        project.install_modules()
+    if strict_deps_check:
+        assert "Incompatibility between constraint y" in e.value.get_message()
+    else:
+        assert "requirements conflicts were found" in e.value.get_message()
