@@ -207,7 +207,7 @@ class LogSequence(object):
             # first error is later
             idxe = self._find("", logging.ERROR, "", self.index)
             assert idxe == -1 or idxe >= index
-        assert index >= 0
+        assert index >= 0, "could not find " + msg
         return LogSequence(self.caplog, index + 1, self.allow_errors, self.ignore)
 
     def assert_not(self, loggerpart, level, msg):
@@ -256,7 +256,7 @@ async def wait_for_version(client, environment, cnt):
         code = compiling.code
         return code == 204
 
-    await retry_limited(compile_done, 10)
+    await retry_limited(compile_done, 30)
 
     reports = await client.get_reports(environment)
     for report in reports.result["reports"]:
